feat(collections): stabilize iterable input for sample
